Move pipeline view to own module
